Perform the shared build in a separate directory on Windows
Perform the shared build first, and in a dedicated build directory
on Windows, to fix a bug where the static build caused the resulting
DLL not to contain any exported symbols.  By doing the shared build
in a separate directory we are ensuring that artifacts from it are not
influencing the static build, and by doing it first we ensure that
artifacts from the static build do not influence the shared build.

I have tried just copying the source directory first, but it caused
more mysterious build failures, and was generally more complex than
this.

Fixes #46

Signed-off-by: Michał Górny <mgorny@quansight.com>